refactor: move postgres functions to database module
